Added PrototypeScene
The PrototypeScene is where I will test mechanics of the game. Currently it has the player that can be controlled with WASD or arrow keys, the speed of the player can be toggled with Shift to simulate sneaking and running speeds. There is also collision boxes on the player, a test tree, and a test rock. All art is currently placeholder.